Split QgsHighlightMaterial to QgsUnlitMaterial base class
Offers a single color, unlit material
